🔧 chore: update .gitignore and package.json dependencies
- Added 'bin' to .gitignore to exclude binary files from version control.
- Included 'fflate' as a new dependency in package.json to support file compression in the application.
- Updated writeFile method in LocalFileCtr to handle file content as Uint8Array for improved type safety.

+++ apps/desktop/src/main/controllers/LocalFileCtr.ts
@@ -646,7 +646,7 @@ export default class LocalFileCtr extends ControllerModule {
         }
 
         await mkdir(path.dirname(targetPath), { recursive: true });
-        await writeFile(targetPath, Buffer.from(fileContent));
+        await writeFile(targetPath, Buffer.from(fileContent as Uint8Array));
       }
 
       await writeFile(markerPath, JSON.stringify({ preparedAt: Date.now(), url, zipHash }), 'utf8');
